Added filter definitions of image with build target
One new file was added: common-targets.yaml. Inside it, the relationship of
images and build targets is defined. It means which image belong to which
target, and the backend services will only generate the proper images for
specific target, to avoid current image duplication issue. It need also cowork
with the new update of `kickstarter` command, and image-configurations build
instruction.

Change-Id: Id3e6ac491b5da059be3f8abc13fa15e551bceb1d
Signed-off-by: Lin Yang <lin.a.yang@intel.com>